third_party/blink/web_tests/virtual/webrtc-wpt-plan-b/external/wpt/webrtc/RTCPeerConnection-operations.https-expected.txt

This is a testharness.js-based test.
FAIL addIceCandidate chains onto SRD, fails before promise_test: Unhandled rejection with value: object "InvalidStateError: Failed to execute 'addTransceiver' on 'RTCPeerConnection': This operation is only supported in 'unified-plan'."
FAIL Pack operations queue with implicit offer and answer promise_test: Unhandled rejection with value: object "InvalidStateError: Failed to execute 'addTransceiver' on 'RTCPeerConnection': This operation is only supported in 'unified-plan'."
FAIL Negotiate solely by operations queue and signaling state promise_test: Unhandled rejection with value: object "InvalidStateError: Failed to execute 'addTransceiver' on 'RTCPeerConnection': This operation is only supported in 'unified-plan'."
PASS promiseState helper works
FAIL Operations chain must run first operation's sync part synchronously assert_equals: expected "InvalidModificationError" but got "OperationError"
FAIL isOperationsChainEmpty detects empty assert_equals: isOperationsChainEmpty expected "InvalidModificationError" but got "OperationError"
FAIL SLD(offer) must validate against LastCreatedOffer early (prerequisite) assert_equals: expected "InvalidModificationError" but got "OperationError"
FAIL SLD(offer) must validate input before signaling state (prerequisite) assert_equals: expected "InvalidModificationError" but got "OperationError"
FAIL createOffer uses operations chain assert_equals: isOperationsChainEmpty expected "InvalidModificationError" but got "OperationError"
FAIL createAnswer uses operations chain assert_equals: isOperationsChainEmpty expected "InvalidModificationError" but got "OperationError"
FAIL setLocalDescription uses operations chain assert_equals: isOperationsChainEmpty expected "InvalidModificationError" but got "OperationError"
FAIL setRemoteDescription uses operations chain assert_equals: isOperationsChainEmpty expected "InvalidModificationError" but got "OperationError"
FAIL addIceCandidate uses operations chain promise_test: Unhandled rejection with value: object "InvalidStateError: Failed to execute 'addTransceiver' on 'RTCPeerConnection': This operation is only supported in 'unified-plan'."
Harness: the test ran to completion.
